requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
diskcache==5.6.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
llama_cpp_python==0.3.14
MarkupSafe==3.0.2
numpy==2.3.2
openai==1.97.0
psutil==7.0.0
puremagic==1.30
pydantic==2.11.7
pydantic_core==2.33.2
pypdf==5.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
pyqterm==0.2
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.17
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pyte==0.8.2
pyttsx3==2.99
shiboken6==6.9.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
watchdog==6.0.0
wcwidth==0.2.13
